feat(GitHub): extract API layer
